Add files via upload
The Main class now has if-else-if statements nested in a while loop to accept user Inputs and return the appropriate message depending on the game of feature that you selected. Since none of the games are available (will add first game tomorrow), all of the options have error messages notifying the user that the game has not been added yet. The quit option allows the user to quit the program. There is an else statement that covers alternate (unintended) user errors at a basic level. The user input is adjusted so that the program will not care about uppercase and lowercase letters. I am already thinking about how I am going to tackle the potential problem of users wanting to play multiple games in one go. I changed the Outputs.java file class name to OutputsInputs.java because I am allowing this class to accept certain user inputs as well. I will figure it out, maybe do another boolean variable with a while loop.