Add files via upload
I added a seperate README document for additional notes, reflections, and documentations. Most of my changes included allowing the user to input which game option that they chose. I did add temporary messages telling the user that the games are not available because I have not added any of the actual games yet. I am planning on uploading the first game tomorrow. I did change the name of the Outputs method to OutputsInputs.

Side note: 
This program really works for me because it allows me to work on smaller gaming programs that I can then just add to this program. This gives me the feeling of making programs without the burnout of really large projects despite the potential for this program to get big.